requirements.txt
pandas==2.2.2
google-cloud-secret-manager==2.20.0
pyspark==3.5.1
google-cloud==0.34.0
google-cloud-bigquery==3.23.1
google-cloud-storage==2.16.0
db-dtypes==1.2.0